refactor(gateway): share connection lifecycle across transports
