+++ config/assertion-safety-baseline.txt
@@ -2689,7 +2689,7 @@ src/gateway/node-command-policy.ts	2
 src/gateway/node-invoke-system-run-approval.ts	1
 src/gateway/node-pairing-ssh-verify.ts	1
 src/gateway/node-registry-private.ts	1
-src/gateway/node-registry.ts	5
+src/gateway/node-registry.ts	4
 src/gateway/openai-agent-run-usage.ts	1
 src/gateway/openai-http.ts	27
 src/gateway/openresponses-http.ts	3
@@ -2816,7 +2816,7 @@ src/gateway/server.auth.modes.suite.ts	1
 src/gateway/server/http-listen.ts	1
 src/gateway/server/plugins-http/route-capability.ts	1
 src/gateway/server/ws-connection-diagnostics.ts	3
-src/gateway/server/ws-connection.ts	3
+src/gateway/server/ws-connection.ts	1
 src/gateway/server/ws-connection/authenticated-request-dispatch.ts	1
 src/gateway/server/ws-connection/connect-device-metadata.ts	1
 src/gateway/server/ws-connection/connect-node-session.ts	1
