+++ src/gateway/role-policy.ts
@@ -1,11 +1,7 @@
 // Gateway connection role policy.
 // Separates node-role RPCs from operator RPCs before method scope checks.
 import { isNodeRoleMethod } from "./method-scopes.js";
-
-const GATEWAY_ROLES = ["operator", "node"] as const;
-
-/** Gateway connection roles used before method-level operator scope checks. */
-export type GatewayRole = (typeof GATEWAY_ROLES)[number];
+import type { GatewayRole } from "./role-policy.types.js";
 
 /** Parses the untrusted role claim from connect params into the closed role set. */
 export function parseGatewayRole(roleRaw: unknown): GatewayRole | null {
