+++ src/gateway/role-policy.types.ts
@@ -1,0 +1,2 @@
+/** Gateway connection roles used before method-level operator scope checks. */
+export type GatewayRole = "operator" | "node";
