+++ src/gateway/server.node-shutdown.test.ts
@@ -1,5 +1,6 @@
 import path from "node:path";
 import { expect, test, vi } from "vitest";
+import { WebSocket } from "ws";
 import { WORKER_EXECUTION_CONTEXT_PROTOCOL_FEATURE } from "../../packages/gateway-protocol/src/schema/worker-admission.js";
 import { runQaGatewayFixture } from "../../test/helpers/qa-gateway-cleanup.js";
 import { writeConfigFile } from "../config/config.js";
@@ -163,6 +164,12 @@ test.for(["direct", "restart"] as const)(
               }
             },
           });
+          const ping = vi.spyOn(WebSocket.prototype, "ping");
+          await expect(
+            kernel.nodeRegistry.checkConnectivity(pairedNode.identity.deviceId),
+          ).resolves.toEqual({ ok: true });
+          expect(ping).toHaveBeenCalledOnce();
+          ping.mockRestore();
           await node.request("node.runnerInventory.update", {
             protocolFeatures: [NODE_WORKER_SUPERVISOR_PROTOCOL_FEATURE],
             workerHost: {
