+++ src/gateway/server/ws-connection/connect-policy.ts
@@ -1,7 +1,7 @@
 // WebSocket connect policy resolves Control UI pairing bypasses and missing-device identity decisions.
 import type { ConnectParams } from "../../../../packages/gateway-protocol/src/index.js";
-import type { GatewayRole } from "../../role-policy.js";
 import { roleCanSkipDeviceIdentity } from "../../role-policy.js";
+import type { GatewayRole } from "../../role-policy.types.js";
 
 export type ControlUiPairingKind = "tailscale-device" | "auth-none" | null;
