+++ src/gateway/server/ws-connection/message-handler.control-ui-build-admission.test.ts
@@ -199,6 +199,7 @@ describe("Control UI build admission over WebSocket", () => {
       };
       attachGatewayWsMessageHandler({
         socket,
+        prepareAuthenticatedReceive: () => ({ ok: true, value: vi.fn() }),
         connectionWork,
         upgradeReq: request as IncomingMessage,
         ingressAttribution: {
