+++ src/gateway/server/ws-connection/message-handler.ts
@@ -1,6 +1,5 @@
 // WebSocket message handler validates frames, dispatches gateway RPCs, manages pairing, and reports responses.
 import { rawDataToString } from "@openclaw/gateway-client/websocket-data";
-import type { RawData } from "ws";
 import {
   GATEWAY_CLIENT_IDS,
   GATEWAY_CLIENT_MODES,
@@ -42,6 +41,7 @@ import { resolveNodePairingClientIpSource } from "../../node-pairing-auto-approv
 import { MAX_PREAUTH_PAYLOAD_BYTES } from "../../server-constants.js";
 import { formatForLog, logWs } from "../../ws-log.js";
 import { truncateCloseReason } from "../close-reason.js";
+import type { GatewayConnectionFrame } from "../connection-transport.js";
 import { resolveGatewayWsBrowserOrigin } from "../ws-origin-policy.js";
 import { createGatewayAuthenticatedRequestDispatcher } from "./authenticated-request-dispatch.js";
 import { isStartupNodeConnect } from "./connect-admission.js";
@@ -54,10 +54,7 @@ import type {
   GatewayConnectPhaseContext,
   GatewayWsMessageHandlerParams,
 } from "./message-handler-types.js";
-export type {
-  GatewayWsMessageHandlerParams,
-  WsOriginCheckMetrics,
-} from "./message-handler-types.js";
+export type { GatewayWsMessageHandlerParams } from "./message-handler-types.js";
 
 const GATEWAY_WORK_ADMISSION_RETRY_AFTER_MS = 1_000;
 const GATEWAY_WORK_ADMISSION_CLOSE_CODE = 1013;
@@ -175,7 +172,7 @@ export function attachGatewayWsMessageHandler(params: GatewayWsMessageHandlerPar
       .catch(onError);
   };
 
-  const handleMessage = async (data: RawData, admission?: "continuation") => {
+  const handleMessage = async (data: GatewayConnectionFrame, admission?: "continuation") => {
     if (isClosed()) {
       return;
     }
@@ -414,7 +411,7 @@ export function attachGatewayWsMessageHandler(params: GatewayWsMessageHandlerPar
   };
 
   const parsePreauthConnectFrame = (
-    data: RawData,
+    data: GatewayConnectionFrame,
   ): { id: string; params: ConnectParams } | null => {
     if (isClosed() || rawDataByteLength(data) > MAX_PREAUTH_PAYLOAD_BYTES) {
       return null;
@@ -435,7 +432,7 @@ export function attachGatewayWsMessageHandler(params: GatewayWsMessageHandlerPar
     return { id: parsed.id, params: parsed.params };
   };
 
-  const isPreparedControlConnect = (data: RawData): boolean => {
+  const isPreparedControlConnect = (data: GatewayConnectionFrame): boolean => {
     const parsed = parsePreauthConnectFrame(data);
     if (!parsed) {
       return false;
@@ -444,12 +441,14 @@ export function attachGatewayWsMessageHandler(params: GatewayWsMessageHandlerPar
     return connectParams.role !== "node" && !claimsWorkerConnectionIdentity(parsed.params);
   };
 
-  const isStartupNodePreauth = (data: RawData): boolean => {
+  const isStartupNodePreauth = (data: GatewayConnectionFrame): boolean => {
     const parsed = parsePreauthConnectFrame(data);
     return parsed ? isStartupNodeConnect(parsed.params) : false;
   };
 
-  const rejectConnectForClosedAdmission = async (data: RawData): Promise<boolean> => {
+  const rejectConnectForClosedAdmission = async (
+    data: GatewayConnectionFrame,
+  ): Promise<boolean> => {
     const parsed = parsePreauthConnectFrame(data);
     if (!parsed) {
       return false;
@@ -487,7 +486,10 @@ export function attachGatewayWsMessageHandler(params: GatewayWsMessageHandlerPar
     return true;
   };
 
-  const handleIncomingMessage = async (data: RawData, requestAdmission?: "continuation") => {
+  const handleIncomingMessage = async (
+    data: GatewayConnectionFrame,
+    requestAdmission?: "continuation",
+  ) => {
     if (getClient()) {
       await handleMessage(data, requestAdmission);
       return;
@@ -537,7 +539,7 @@ export function attachGatewayWsMessageHandler(params: GatewayWsMessageHandlerPar
     }
   };
 
-  socket.on("message", (data) => {
+  const receive = (data: GatewayConnectionFrame) => {
     // Capture receipt before any await: older requests keep their admitted lifetime,
     // while shutdown frames may only settle an exact pending node owner.
     const admission = params.connectionWork.isClosing ? "continuation" : undefined;
@@ -553,5 +555,7 @@ export function attachGatewayWsMessageHandler(params: GatewayWsMessageHandlerPar
       .catch((error: unknown) => {
         logGateway.error(`request dispatch failed conn=${connId}: ${formatForLog(error)}`);
       });
-  });
+  };
+  socket.on("message", receive);
+  return receive;
 }
