+++ src/gateway/server/ws-connection/request-start.ts
@@ -1,64 +1,7 @@
 import { performance } from "node:perf_hooks";
 import { setImmediate as nextTurn } from "node:timers/promises";
-import type { WebSocket } from "ws";
 import { runOutsideGatewayRootWorkAdmission } from "../../../process/gateway-work-admission.js";
 import { BoundedSerialQueue } from "../../../shared/bounded-serial-queue.js";
-import type { GatewayRole } from "../../role-policy.js";
-import { MAX_PAYLOAD_BYTES } from "../../server-constants.js";
-
-type PayloadLimited = { _maxPayload?: number };
-type GatewayReceiver = PayloadLimited & {
-  _allowSynchronousEvents?: boolean;
-};
-
-function hasWritablePayloadLimit(target: PayloadLimited | undefined): target is PayloadLimited {
-  return (
-    typeof target?.["_maxPayload"] === "number" &&
-    Object.getOwnPropertyDescriptor(target, "_maxPayload")?.writable === true
-  );
-}
-
-function gatewayReceiver(socket: WebSocket): GatewayReceiver | null {
-  // SAFETY: ws owns these private per-frame fields; validate each before the handoff.
-  const receiver = (socket as WebSocket & { _receiver?: GatewayReceiver })["_receiver"];
-  if (!hasWritablePayloadLimit(receiver)) {
-    return null;
-  }
-  return receiver;
-}
-
-/** Raises the authenticated frame limit after the connection is admitted. */
-export function raiseGatewayReceiverPayloadLimit(socket: WebSocket, maxPayload: number): boolean {
-  const receiver = gatewayReceiver(socket);
-  if (!receiver) {
-    return false;
-  }
-  receiver["_maxPayload"] = maxPayload;
-  return true;
-}
-
-export function prepareGatewayReceiverHandoff(
-  socket: WebSocket,
-  role: GatewayRole,
-): (() => void) | null {
-  const receiver = gatewayReceiver(socket);
-  if (!receiver) {
-    return null;
-  }
-  if (
-    role === "operator" &&
-    (typeof receiver["_allowSynchronousEvents"] !== "boolean" ||
-      Object.getOwnPropertyDescriptor(receiver, "_allowSynchronousEvents")?.writable !== true)
-  ) {
-    return null;
-  }
-  return () => {
-    receiver["_maxPayload"] = MAX_PAYLOAD_BYTES;
-    if (role === "operator") {
-      receiver["_allowSynchronousEvents"] = true;
-    }
-  };
-}
 
 // One active scheduling task is separate from these waiting budgets. Each task
 // grants start permission only; it never owns the RPC or waits for its completion.
